+++ base/profile/src/main/java/io/snaps/baseprofile/domain/BalanceModel.kt
@@ -1,7 +1,7 @@
 package io.snaps.baseprofile.domain
 
 data class BalanceModel(
-    val unlocked: Int,
-    val locked: Int,
+    val unlocked: Double,
+    val locked: Double,
     val exchangeRate: Double,
 )
